Map sidebar reorganization
