+++ frontend/src/containers/map/sidebar/layers-panel/index.tsx
@@ -1,0 +1,5 @@
+const LayersPanel: React.FC = () => {
+  return <div>LAYERS</div>;
+};
+
+export default LayersPanel;
