--- frontend/src/containers/map/sidebar/layers/index.tsx
@@ -1,5 +1,0 @@
-const SidebarLayers: React.FC = () => {
-  return <div>LAYERS</div>;
-};
-
-export default SidebarLayers;
