+++ frontend/src/containers/map/sidebar/main-panel/details/details-button/index.tsx
