+++ frontend/src/containers/map/sidebar/main-panel/details/index.tsx
