+++ frontend/src/containers/map/sidebar/main-panel/details/location-selector/location-dropdown/index.tsx
