+++ frontend/src/containers/map/sidebar/main-panel/details/location-selector/type-toggle/index.tsx
