+++ frontend/src/containers/map/sidebar/main-panel/details/widgets/establishment-stages/index.tsx
