+++ frontend/src/containers/map/sidebar/main-panel/details/widgets/habitat/index.tsx
