+++ frontend/src/containers/map/sidebar/main-panel/details/widgets/index.tsx
