+++ frontend/src/containers/map/sidebar/main-panel/details/widgets/marine-conservation/index.tsx
