+++ frontend/src/containers/map/sidebar/main-panel/details/widgets/protection-types/index.tsx
