+++ frontend/src/containers/map/sidebar/main-panel/modelling/modelling-intro/index.tsx
